+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.2.0] - 2026-04-28
+
+### Added ###
+
+- Added option screen support
+- Added better control remap support
+- Added a static event (OnThemeChanged) for when the current theme has changed (for mod developpers)
+
 ## [0.1.0] - 2026-04-27
 
 - Beta Release